requirements.txt
apache-airflow==3.0.2
pendulum==3.1.0
pytest==8.4.1
Requests==2.32.4
apache-airflow-providers-http
